requirements.txt
streamlit
PyMuPDF
openai
pandas
python-docx
pytesseract
langchain_text_splitters
pypandoc